\name{BabbleClient-package}
\alias{BabbleClient-package}
\alias{BabbleClient}
\docType{package}
\title{BabbleClient:  Convert Biological Identifiers}
\description{
Provides programmatic access to ISB's gdxbase identifier conversions
}
\details{
\tabular{ll}{
Package: \tab BabbleClient\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-09\cr
License: \tab LGPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Victor Cassen <vcassen@systemsbiology.org>
Maintainer: Victor Cassen <vcassen@systemsbiology.org>
}
\references{
none yet...
}
\keyword{Annotation}
\seealso{
  NA
}
\examples{
  NA
}
