\name{translate}
\alias{translate}
\title{translate}
\description{
translate the input_ids, of input_type, to output_ids, of the output type.
}
\usage{
translate (input_type,input_ids,output_types)
}
\arguments{
  \item{input_type}{a \code{string} object, one of the permitted values
    (from idtypes) }
  \item{input_ids}{a list of \code{strings} object, typically from 0 to 5.}
  \item{output_types}{a \code{string} object, one of the permitted values
    (from idtypes) }
}

\value{
list of strings
}
\author{Victor Cassen}


\examples{
  translate ()
}

\keyword{annotation}
